refactor: ChatRoom, ChatMessage 도메인 분리 #7
