+++ src/main/java/com/example/chatserver/domain/chatMessage/dto/payload/ChatMessagePayload.java
@@ -1,0 +1,18 @@
+package com.example.chatserver.domain.chatMessage.dto.payload;
+
+import com.example.chatserver.common.entity.ChatMessage;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ChatMessagePayload {
+
+    private Long roomId;
+    private Long senderId;
+    private String content;
+
+    public static ChatMessagePayload from(ChatMessage message) {
+        return new ChatMessagePayload(message.getChatRoom().getId(), message.getSenderId(), message.getContent());
+    }
+}
