--- src/main/java/com/example/chatserver/dto/request/SendMessageRequest.java
@@ -1,10 +1,0 @@
-package com.example.chatserver.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class SendMessageRequest {
-
-    private Long receiverId;
-    private String content;
-}
