--- src/main/java/com/example/chatserver/dto/response/OpenChatRoomResponse.java
@@ -1,16 +1,0 @@
-package com.example.chatserver.dto.response;
-
-import com.example.chatserver.entity.ChatRoom;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OpenChatRoomResponse {
-
-    private Long roomId;
-
-    public static OpenChatRoomResponse from(ChatRoom chatRoom) {
-        return new OpenChatRoomResponse(chatRoom.getId());
-    }
-}
