--- src/main/java/com/example/chatserver/repository/MessageRepository.java
@@ -1,7 +1,0 @@
-package com.example.chatserver.repository;
-
-import com.example.chatserver.entity.ChatMessage;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MessageRepository extends JpaRepository<ChatMessage,Long> {
-}
