+++ src/main/java/com/example/chatserver/test/dto/request/SendMessageRequest.java
@@ -1,0 +1,10 @@
+package com.example.chatserver.test.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class SendMessageRequest {
+
+    private Long receiverId;
+    private String content;
+}
