+++ src/main/java/com/example/chatserver/test/dto/response/SendMessageResponse.java
@@ -1,0 +1,19 @@
+package com.example.chatserver.test.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@AllArgsConstructor
+public class SendMessageResponse {
+
+    private Long senderId;
+    private String content;
+    private LocalDateTime sentAt;
+
+    public static SendMessageResponse of(Long senderId, String content) {
+        return new SendMessageResponse(senderId, content, LocalDateTime.now());
+    }
+}
