--- src/main/java/com/example/chatserver/testUser/StompPrincipal.java
@@ -1,15 +1,0 @@
-package com.example.chatserver.testUser;
-
-import lombok.Getter;
-
-import java.security.Principal;
-
-@Getter
-public class StompPrincipal implements Principal {
-
-    private final String name;
-
-    public StompPrincipal(String name) {
-        this.name = name;
-    }
-}
